chore(FancyTable): Move to common
